Se modifican servicios de eliminar hots, puestos y licensias para añadirle la validación de que no pueden estar asociados a un proyecto para poder ser eliminados

+++ backend/lib/application/use_cases/init_project/delete_db_hosts.js
@@ -1,11 +1,16 @@
 "use strict";
 
-module.exports = async (id, { gnrDbHostsRepository }) => {
+module.exports = async (
+  id,
+  { gnrDbHostsRepository, gnrProjectsRepository }
+) => {
   const dbHosts = await gnrDbHostsRepository.getByFilter({ id });
   if (dbHosts === null) return 432;
   if (!dbHosts) return false;
 
-  //Falta verificar que este host no se encuentre asociado a un proyecto
+  //Verificar que este host no se encuentre asociado a un proyecto
+  const project = await gnrProjectsRepository.getByFilter({ dbHostId: id });
+  if (project) return 441;
 
   return await gnrDbHostsRepository.remove(id);
 };
