+++ backend/lib/application/use_cases/init_project/delete_db_port.js
@@ -1,15 +1,13 @@
 "use strict";
 
-module.exports = async (id, { gnrDbPortsRepository, gnrD }) => {
+module.exports = async (id, { gnrDbPortsRepository, gnrProjectsRepository }) => {
   const dbPort = await gnrDbPortsRepository.getByFilter({ id });
   if (dbPort === null) return 432;
   if (!dbPort) return false;
 
   //Verificar que este puerto no se encuentre asociado a un proyecto
-//   const project = await fuCategoryItemRepository.getByFilter({
-//     itemId: id,
-//   });
-//   if (project) return 441;
+  const project = await gnrProjectsRepository.getByFilter({ dbPortId: id })
+  if (project) return 441;
 
   return await gnrDbPortsRepository.remove(id);
 };
